Move copied `finalize_execution`
